requirements.txt
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==0.28
altair==4
streamlit==1.24.1
langchain==0.0.330
codeinterpreterapi[all]
pickle-mixin

